+++ frontend/src/app/Components/MainPages/shorten-url-page/shorten-url-page.component.html
@@ -60,25 +60,27 @@
       </div>
     </div>
 
-    <div>
-      <button type="button" id = "tagCreate" (click)="openCreateTagView()">Create New Tag</button><br>
-      <label for="tagInput">Add tags </label>
-      <input type="text" id = "tagInput" placeholder="tag" formControlName="tagInput" required list="suggestions" autocomplete="off"/>
-      <datalist id="suggestions">
-        <option *ngFor="let tag of availableTags" [value]="tag.title"></option>
-      </datalist>
-      <button type="button" id = "tagAdd" (click)="addTag()" [disabled]="getTagAddDisabled()">Add</button><br>
+    <div id="tagSection">
+      <label id="tagsTitle" for="tagInput">Add tags </label>
+      <div>
+        <input type="text" id="tagInput" placeholder="tag" formControlName="tagInput" required list="suggestions" autocomplete="off"/>
+        <datalist id="suggestions">
+          <option *ngFor="let tag of availableTags" [value]="tag.title"></option>
+        </datalist>
+        <button type="button" id="tagAdd" (click)="addTag()" [disabled]="getTagAddDisabled()">Add</button>
+        <button type="button" id="tagCreate" (click)="openCreateTagView()">New Tag</button>
+      </div>
       <span *ngFor="let tag of currentTags">
-        <span class="tagShell" style="background-color:{{tag.color}};">
-          {{tag.title}}
-        </span>
-        <button type="button" class="tagDelete" (click)="deleteTag(tag)"><b>X</b></button>
+      <span class="tagShell" [style]="{'background': tag.color, 'color': isTagColorLight(tag.color)? 'black': 'white'}">
+        {{tag.title}}
       </span>
+      <button type="button" class="tagDelete" (click)="deleteTag(tag)"><b>X</b></button>
+    </span>
     </div>
 
     <br>
     <div class="horizCenter bottomSpace">
-      <input type="submit" class="button" value="Shorten URL" />
+      <input id="submit" type="submit" class="button" value="Shorten URL" />
     </div>
   </form>
   </div>
